resolve componentes and make a small documentation
